Apply change from upstream
https://github.com/qt/qtbase/commit/e7efa77bb3e730c123411f3ed679036df812f57f

+++ src/gui/widgets/qvalidator.cpp
@@ -606,7 +606,7 @@ QValidator::State QDoubleValidatorPrivate::validateWithLocale(QString &input, QL
 
     if (notation == QDoubleValidator::StandardNotation) {
         double max = std::max(std::abs(q->b), std::abs(q->t));
-        if (max < LLONG_MAX) {
+        if (max < double(LLONG_MAX)) {
             qlonglong n = pow10(numDigits(qlonglong(max))) - 1;
             if (std::abs(i) > n)
                 return QValidator::Invalid;
